+++ packages/components-js/tests/vrt/playwright/helpers/popover-helper.ts
@@ -1,0 +1,10 @@
+import { type Page } from '@playwright/test';
+
+export const openAllPopover = async (page: Page): Promise<void> => {
+  await page.evaluate(() => {
+    document.addEventListener('mousedown', (e) => e.stopPropagation(), true);
+    document.querySelectorAll('p-popover, my-prefix-p-popover').forEach((popover) => {
+      popover.shadowRoot.querySelector('button').click();
+    });
+  });
+};