+++ packages/components-js/tests/vrt/playwright/specs/states/display.vrt.ts
@@ -1,0 +1,62 @@
+import { expect, type Page, test } from '@playwright/test';
+import {
+  baseSchemes,
+  baseThemes,
+  baseViewportWidth,
+  forceFocusHoverState,
+  forceFocusState,
+  forceHoverState,
+  getPlaygroundPseudoStatesMarkup,
+  type PrefersColorScheme,
+  setContentWithDesignSystem,
+} from '../../helpers';
+import { type Theme } from '@porsche-design-system/utilities-v2';
+
+const component = 'display';
+
+const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): Promise<void> => {
+  const markup = () => `
+    <p-display size="medium">
+      Display
+      <span>
+        and some slotted, deeply nested <a href="#">anchor</a>.
+      </span>
+    </p-display>`;
+
+  await setContentWithDesignSystem(page, getPlaygroundPseudoStatesMarkup(markup), {
+    forceComponentTheme: theme,
+    prefersColorScheme: scheme,
+  });
+
+  await forceHoverState(page, '.hover p-display a');
+  await forceHoverState(page, '.hover p-display button');
+  await forceFocusState(page, '.focus p-display a');
+  await forceFocusState(page, '.focus p-display button');
+  await forceFocusHoverState(page, '.focus-hover p-display a');
+  await forceFocusHoverState(page, '.focus-hover p-display button');
+};
+
+// executed in Chrome only
+test.describe(component, async () => {
+  test.skip(({ browserName }) => browserName !== 'chromium');
+
+  baseThemes.forEach((theme) => {
+    test(`should have no visual regression for :hover + :focus-visible with theme ${theme}`, async ({ page }) => {
+      await scenario(page, theme);
+      await expect(page.locator('#app')).toHaveScreenshot(
+        `${component}-${baseViewportWidth}-states-theme-${theme}.png`
+      );
+    });
+  });
+
+  baseSchemes.forEach((scheme) => {
+    test(`should have no visual regression for :hover + :focus-visible with theme auto and prefers-color-scheme ${scheme}`, async ({
+      page,
+    }) => {
+      await scenario(page, 'auto', scheme);
+      await expect(page.locator('#app')).toHaveScreenshot(
+        `${component}-${baseViewportWidth}-states-theme-${scheme}.png`
+      ); // fixture is aliased since result has to be equal
+    });
+  });
+});