+++ packages/components-react/projects/nextjs/routes.ts
@@ -128,10 +128,6 @@ const generatedRoutes = {
     "path": "/overview",
     "name": "Overview"
   },
-  "overviewFlaky": {
-    "path": "/overview-flaky",
-    "name": "OverviewFlaky"
-  },
   "overviewNotifications": {
     "path": "/overview-notifications",
     "name": "OverviewNotifications"
@@ -172,6 +168,10 @@ const generatedRoutes = {
     "path": "/switch",
     "name": "Switch"
   },
+  "table": {
+    "path": "/table",
+    "name": "Table"
+  },
   "tabs": {
     "path": "/tabs",
     "name": "Tabs"
@@ -204,9 +204,9 @@ const generatedRoutes = {
     "path": "/textarea-wrapper",
     "name": "TextareaWrapper"
   },
-  "toastBasic": {
-    "path": "/toast-basic",
-    "name": "ToastBasic"
+  "toast": {
+    "path": "/toast",
+    "name": "Toast"
   },
   "toastBasicDark": {
     "path": "/toast-basic-dark",
@@ -232,9 +232,9 @@ const generatedRoutes = {
     "path": "/typography-cyril",
     "name": "TypographyCyril"
   },
-  "typographyFallbackStrategy": {
-    "path": "/typography-fallback-strategy",
-    "name": "TypographyFallbackStrategy"
+  "typographyFallback": {
+    "path": "/typography-fallback",
+    "name": "TypographyFallback"
   },
   "typographyGreekAndCoptic": {
     "path": "/typography-greek-and-coptic",
@@ -253,7 +253,6 @@ const generatedRoutes = {
 
 export const sitemap = {
   home: { path: '/', name: 'Home' },
-  table: { path: '/table', name: 'Table' },
   ...generatedRoutes,
 };