+++ packages/components-vue/tests/vrt/config/jest.setup.ts
@@ -1,7 +1,0 @@
-import { setCustomOptions } from '@porsche-design-system/shared/testing';
-
-setCustomOptions({
-  baseUrl: 'http://localhost:5173',
-  fixturesDir: '../components-js/tests/vrt/puppeteer/fixtures',
-  resultsDir: 'tests/vrt/results',
-});