+++ packages/components/src/components/toast/toast-item/toast-item-styles.ts
@@ -7,7 +7,7 @@ import {
   getNotificationRootJssStyle,
 } from '../../inline-notification/inline-notification-styles-shared';
 import { dropShadowHighStyle, textSmallStyle } from '@porsche-design-system/utilities-v2';
-import { getThemedColors } from '../../../styles';
+import { getThemedColors, prefersColorSchemeDarkMediaQuery } from '../../../styles';
 
 export const getComponentCss = (state: ToastState, theme: Theme): string => {
   return getCss({
@@ -20,6 +20,9 @@ export const getComponentCss = (state: ToastState, theme: Theme): string => {
         ...textSmallStyle,
         margin: 0,
         color: getThemedColors(theme).primaryColor,
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: getThemedColors('dark').primaryColor,
+        }),
       },
     },
     icon: getNotificationIconJssStyle(),